Implementation of audio
Near completed implmentation, featuring:
Completely debugged and safe (for DEFAULT/STANDARD CANONICAL CODEC ONLY) WAV file loading
Implemented various parameters for audio. In terms of user available parameters, we have pitch and audio.
The audio manager is now implemented as a singleton. Channels are managed as arrays of sources - thus, there is a limited amount of sounds we can play at once, but we can configure that as needed depending on necessary performance.
There are a variety of quirky possible(?)  small performance concerns we can address later.
The sound component as of now only implements playing the current audio, yet the other methods are trivial to implement (I am tired as fuck rn kek)
Looping for now will just be playing back the current source over and over.
Implemented two enums, one for letting users decide priority of the sound components they create, and the other for getting the current state of a sound component (not yet implemented)